requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests

langchain
langgraph
google-generativeai

faiss-cpu
sentence-transformers

psycopg2-binary
sqlalchemy

python-telegram-bot

apscheduler

pytest

pymysql

faiss-cpu
sentence-transformers